Extracting progress reporter to own class
